Add agents and maze environment for reinforcement learning
This commit introduces a new `agents.py` module containing three agent classes: `SimpleAgent`, `MemoryAgent`, and `RandomAgent`, designed for reinforcement learning tasks. The `MemoryAgent` enhances Q-learning with episodic memory using a `MemorySpace`. Additionally, a new `maze.py` module is added, providing a `MazeEnvironment` class that simulates a grid-based maze for agent navigation. The `main_demo.py` file is updated to utilize these new classes, facilitating experiments with memory-enabled and memory-disabled agents. Utility functions for converting NumPy types to Python types are also included in a new `util.py` module for JSON serialization.
